chore(plan-strategy-refactor): auto-commit after run-summary session
