+++ src/plan/strategies/factory.ts
@@ -1,0 +1,21 @@
+import { NaxError } from "../../errors";
+import type { IPlanStrategy } from "./types";
+import { DebatePlanStrategy } from "./debate";
+import { PipelinePlanStrategy } from "./pipeline";
+import { SinglePlanStrategy } from "./single";
+
+export function createPlanStrategy(mode: IPlanStrategy["mode"]): IPlanStrategy {
+  switch (mode) {
+    case "single":
+      return new SinglePlanStrategy();
+    case "pipeline":
+      return new PipelinePlanStrategy();
+    case "debate":
+      return new DebatePlanStrategy();
+    default:
+      throw new NaxError(`[plan] Unknown plan mode: ${mode}`, "PLAN_MODE_UNKNOWN", {
+        stage: "plan",
+        mode,
+      });
+  }
+}
